+++ src/shared/index.ts
@@ -2,4 +2,5 @@ export * from "./assets";
 export * from "./components";
 export * from "./hooks";
 export * from "./sections";
+export * from "./types";
 export * from "./utils";
